feat($search): moved to Postlist (in refact)
